sample/gae/requirements.txt
genpac-server